Correct description of v3 introduce cells to match implementation
We said that every v3 introduce cell needs to have an AUTHL field.
But it turns out we only produce that field when we are using a
nonzero authentication method.  It's not possible to make clients or
services begin matching the spec without incompatibility, so instead
let's make the spec match our behavior.