Switching ConnType to be an enum
I'm still not really sure why Karsten wants to use a string field for this.
This has all the hallmarks of being an enumeration and it's trivially easy to
add new connection types as they arise.

Using a string for this field is not just unusual, but makes things a bit
harder for both controller users and anyone trying to build an alternative tor
implementation.

I'm also swapping the enums to be uppercase since that's the present convention
in the control-spec.