Noting that GETINFO ns/* options provide v3 directory information
Tor moved from v2 to v3 directory information in commit 3ad6dc0e. Prior to that
Tor's 'GETINFO ns/*' options provided v2 information, but since then it's been
v3. For more information see...

https://trac.torproject.org/7953

Technically this is a tor bug, but it happens to be a convenient one so simply
updating our spec. In practice controller authors haven't noticed this
discrepancy because v3 router status entries are a superset of the fields in
v2.