Make new subsection for routers uploading descriptors.
The rule we're trying to follow here is that sections are for roles
(client, authority, cache, client) and subsections are for operations
performed by the role.  Routers really only perform a single
operation: they periodically upload their descriptors to the
authorities.  That's why there should be a single subsection for this
operation.

This effectively substitutes section number parts "2." with "2.1.":
 * 2 -> 2.1
 * 2.1 -> 2.1.1
 * 2.2 -> 2.1.2
 * 2.2.1 -> 2.1.2.1
 * 2.3 -> 2.1.3