Make section references easier to search for.
Prefix all section references with the word "section", and avoid line
breaks between "section[s]" and section numbers.

Searching for section references should now be as easy as (for
example):

  grep "section.* 1\.4" dir-spec.txt

Also, while going through section references, lower-case 1 instance of
"Section" to be consistent with the rest and fix 2 broken references.