If the consensus doesn't contain exits, don't build exit paths
If the consensus doesn't contain exits, we only build internal paths.
This is enough to allow reachability tests (which can enable exits to
bootstrap), and hidden services.

If we subsequently receive a consensus with exits, start building exit paths.

Update dir-spec and path-spec to document this.
Update control-spec to document changes in controller bootstrap messages.

Based on changes made in tor to resolve bug #13814.