Rewrite the existing PT spec to be not Tor specific.
This attempts to make it clear that PTs are not just for Tor, and can
be used by any project, and should be sufficient documentation for
writing the PT glue code both for Tor and other projects.

TODO: Fold in the implemented parts of prop 196/217.  I'll do this when
I have time, since statistics are useful for everybody.

Fixes: #13369, #15545
Completes: #16754