prop224: Revisit basic client authorization
In the past prop224 used to embed the client authorization key in the
subcredential. The problem here is that if we wanted to revoke a client,
we would have to change the whole subcredential of the service, which
means that we would have to announce it to all clients.

This patch makes it so that every client has an x25519 and an ed25519
which are used to perform client authorization.

To achieve this on the descriptor level, we change the descriptor format
to a double-layer encryption where the first layer protects against
entities who don't know the public key of the HS, and the second layer
protects against unauthorized clients who don't know the x25519 key.

The intro level authorization remains as is and uses ed25519 for authentication.

Thanks to special for noticing this issue.

Thanks to Nick for sketching out the x25519 descriptor auth scheme.