Change header ordering from a MUST to SHOULD clause
We had this as a 'MUST' because the directory authorities need to agree upon an
ordering in their votes to derive at a signed consensus. However, this
requirement is turning out to be very error prone...

  https://trac.torproject.org/projects/tor/ticket/21059

Got the ordering issues sorted out... I think. However, this has bitten us
three times in a row and clearly will bite us more in the future so loosening
these requirements a little.