Fix a security problem in prop140 related to signature encoding
The problem was that clients would, when contacting caches, identify
consensuses by the sha3 digest of the entire consensus, including
signatures.  But there are multiple valid encodings for a set of
signatures, meaning that a malicious cache could serve each client a
different encoding, and recognize the clients using the sha3 digests
in their requests.

The first part of the solution is to fetch consensuses diffs based
only on the consensus's digest-as-signed: the digest of the
consensus with no signatures on it.

The second part of the solution is to generate diffs using the
<n>,$d format to first remove all trailing signatures, so that the
diffs will apply to any valid consensus, no matter how the
signatures are encoded.