minor touch-ups on the glossary
the only exciting one is that we don't use guards to defend against
"enumeration attacks" -- i'm not quite sure what an enumeration attack
is, but it sounds like something where the guard is able to make a list
of users, and where having that list is bad news in itself. that's not
quite what guards are for.