Drop 'version added' clause from 'GETINFO sr/current'
This clause was added in...

  https://gitweb.torproject.org/torspec.git/commit/?id=5c82d5e

Commit message and ticket make no mention of 'GETINFO sr/current' go I suspect
Daniel meant for this to concern his addition. However, if that's the case then
according to the ticket it should be 0.3.1.1-alpha.

Just gonna drop this clause since it's probably wrong.