Clarify the TAP hybrid encryption scheme
This is only used in TAP and old-style hidden services, and it's
half malleable. I've clarified how the code behaves by adding the
change suggested in #22987.  I've also noted:

I've also noted that we don't actually reach case 1 with any usage
of this algorithm.

I've also replaced Roger's note that someday we'll add a MAC with
an admonition not to use this hybrid encryption approach for
anything new.  We're not planning to add a MAC; we've migrated to
ntor instead.