prop224: Specify new descriptor upload/fetching behavior.
As part of our work in #23387, we figured out that there are some edge
cases where clients cannot connect to services if they are using
different live consensuses. That was because the overlap period was only
covering clients with a newer consensus than the service.

We are now extending the overlap period to be permanent, and alter its
logic so that it also covers clients with older consensus than the
service.

Now services always have two active descriptors at any given time.

This spec patch is a companion to the code branch of #23387.