Require a "Ticket:" field for future proposals in certain states.
New proposals in states {OPEN,ACCEPTED,CLOSED,FINISHED} must have a
"Ticket:" field which may either be a ticket number or a publicly
accessible URL.  When the ticket is just a number, it should be in
reference to Tor's canonical bug tracker.

This information greatly helps both new contributors and relay
operators who are interested in the implementation status of a working
proposal.  This patch permits the "Ticket:" field to be added to older
proposals (<= prop#288), but it does not require it.