Attempt to clarify new CIRC_BW fields
I didn't really follow the last description so trying one of my own...

  https://trac.torproject.org/projects/tor/ticket/26110

Note that this description likely is *not* accurate. I didn't dig through tor's
source to puzzle out what these fields are. Rather, this is a more grokable
description of what I *think* the fields are. On the above ticket the Network
team and I will work together on a description that's both easily understood
and accurate.