dir-spec: The ns consensus isn't explicitly ns
tor does not generate ns consensuses that are explicitly labelled as ns
consensuses. This updates the spec to make the flavor portion optional,
and to allow the assumption to be made that if a flavor is omitted from
the "network-status-version" line then it is "ns".

One "flavour" was also turned into a "flavor" for consistency.