prop289: Rewrite the Deployment Plan section
Split it into subsection for clarity. Add a new subsection describing the
addition of a new protover value.

This patch removes a paragraph that states the obvious with "We'll need to
do a bunch of testing". It is nice but also huge part of our development
work flow so no need for it in the proposal.

Signed-off-by: David Goulet <dgoulet@torproject.org>