bandwidth-file: Add relay line diagnostic KeyValues
Make Tor directory authorities avoid voting on relay bandwidths
for relays that were not successfully measured. But report diagnostics
for those relays in the Bandwidth File.

When there are not enough eligible relays, put all the relays in the file
for diagnostic purposes. But mark them so that Tor avoids voting on their
bandwidths.

Closes 29813, 29853.