build HTML GitLab Pages site with GitLab CI
This does a rough conversion to Markdown using regexps, then uses pandoc to
convert the Markdown to HTML.  An index.html is also generated to make it
easy to find the various documents.  This will create a site in any GitLab
fork that has CI/CD and Pages setup, which is the default on gitlab.com.