bandwidth-file: Add KeyValues to version 1.4.0
They appeared in in bandwidth files version 1.4.0 generated by bwauths,
but they were forgotten when writing this specification.
Even if the KeyValues related to bandwidth values were added before
the monitoring KeyValues, there was not a sbws release nor a bwauth
that published those KeyValues with a specification version previous
to 1.4.0.
They have not been added to the A.5 section because those Bandwidth
Lines do not contain all the KeyValues.

Closes #30726.