Fix some descriptions in cert_key_type_fix.
There were extraneous entries in the lists of signed key types:
these values describe the type of the key being signed (eg, ed25519
or RSA), not the type of the certificate.

Further, clarify Tor bug 40124, so that people know to expect
incorrect cert key types on older signing->link certs.