Better description for when a consensus must is "too early"
Since the authorities can produce a signed consensus as soon as
`ValidAfter` minus `DistSeconds`, and since they serve a signed
consensus as soon as it has enough signatures, it's possible that a
client or relay that's starting late in the hour can get an "early"
consensus.  Back in tor#25756, we fixed this issue in Tor, but we
didn't document the behavior in the spec.