prop#324: mention rearranged form for N_EWMA smoothing
The formula used for N_EWMA is actually rearranged in the C tor
implementation, in a way that caused the arti reimplementation to be off
by a small rounding error until it was corrected.

In order to ensure other implementations don't have this issue, mention
the rearranged form and add a requirement that it be used.

(background:
https://gitlab.torproject.org/tpo/core/arti/-/merge_requests/525#note_2807244)